requirements.txt
setuptools
tequila-basic
